added new features to the ATOC algos

+++ src/RandomStuff/FermatNumbers.java
@@ -4,8 +4,8 @@ import Utils.ConsolePrinting;
 
 public class FermatNumbers {
 
-    public static int fermatNumber(int p) {
-        return (int) Math.pow(2, Math.pow(2, p)) + 1;
+    public static long fermatNumber(int p) {
+        return (long) Math.pow(2, Math.pow(2, p)) + 1;
     }
     public static void main(String[] args) {
         for(int i = 0; i < 1000; i++) {
